Removed the `FastCacheCore` class
As far as I could see, only the `ExcludedPaths` field was being used, which itself was a wrapper for `Configuration.ExcludePaths`, (and only used in one place).
So I referenced the `Configuration.ExcludePaths` in the module code and removed the rest of this class.

The `ClearCache` and `GetCacheFiles` methods weren't used elsewhere in the codebase.